requirements-ordenes.txt
PyDispatcher==2.0.7
SQLAlchemy==1.4.44
aiopulsar-py==0.1.0a0
fastapi==0.89.1
pulsar-client==2.10.2
pulsar-client[avro]==2.10.2
requests==2.28.2
uvicorn[standard]==0.20.0
